style: 잘못된 클래스 이름 수정 및 메서드 분리
